+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Client/ContractTypeStringType.cs
@@ -1,0 +1,7 @@
+﻿namespace Vodovoz.Domain.Client
+{
+	public class ContractTypeStringType : NHibernate.Type.EnumStringType
+	{
+		public ContractTypeStringType() : base(typeof(ContractType)) { }
+	}
+}
