+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Client/PaymentTypeStringType.cs
@@ -1,0 +1,7 @@
+﻿namespace Vodovoz.Domain.Client
+{
+	public class PaymentTypeStringType : NHibernate.Type.EnumStringType
+	{
+		public PaymentTypeStringType() : base(typeof(PaymentType)) { }
+	}
+}
