+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Goods/VATStringType.cs
@@ -1,0 +1,7 @@
+﻿namespace Vodovoz.Domain.Goods
+{
+	public class VATStringType : NHibernate.Type.EnumStringType
+	{
+		public VATStringType() : base(typeof(VAT)) { }
+	}
+}
