+++ Source/Libraries/Core/Business/VodovozBusiness/VodovozBusiness.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -53,6 +53,8 @@
   <ItemGroup>
     <Compile Include="CommonEnums\Month.cs" />
     <Compile Include="CommonEnums\AllYesNo.cs" />
+    <Compile Include="Domain\Client\ContractType.cs" />
+    <Compile Include="Domain\Client\ContractTypeStringType.cs" />
     <Compile Include="Domain\Goods\AlternativeNomenclaturePrice.cs" />
     <Compile Include="Domain\Goods\NomenclaturePriceBase.cs" />
     <Compile Include="Domain\Logistic\AddressTransferType.cs" />
@@ -205,6 +207,7 @@
     <Compile Include="Domain\Orders\TrueMarkApiDocument.cs" />
     <Compile Include="Domain\Organizations\OrganizationOwnershipType.cs" />
     <Compile Include="Domain\Organizations\OrganizationVersion.cs" />
+    <Compile Include="Domain\Client\PaymentTypeStringType.cs" />
     <Compile Include="Domain\Permissions\AvailableDatabase.cs" />
     <Compile Include="Domain\Permissions\DatabasePrivilege.cs" />
     <Compile Include="Domain\Permissions\GlobalPrivilege.cs" />
@@ -286,6 +289,8 @@
     <Compile Include="Domain\TrueMark\CashReceiptStatus.cs" />
     <Compile Include="Domain\TrueMark\TrueMarkCashReceiptProductCode.cs" />
     <Compile Include="Domain\TrueMark\TrueMarkIdentificationCode.cs" />
+    <Compile Include="Domain\Goods\VAT.cs" />
+    <Compile Include="Domain\Goods\VATStringType.cs" />
     <Compile Include="Domain\Versions\VersionEntityBase.cs" />
     <Compile Include="Domain\WageCalculation\CalculationServices\RouteList\RouteListItemWageCalculationDetails.cs" />
     <Compile Include="Domain\WageCalculation\CalculationServices\RouteList\RouteListWageCalculationService.cs" />
@@ -755,7 +760,7 @@
     <Compile Include="Domain\Client\RepairAgreement.cs" />
     <Compile Include="Domain\Client\WaterSalesAgreement.cs" />
     <Compile Include="Domain\CommentTemplate.cs" />
-    <Compile Include="Domain\CommonEnums.cs" />
+    <Compile Include="Domain\Client\PaymentType.cs" />
     <Compile Include="Domain\Documents\CarLoadDocument.cs" />
     <Compile Include="Domain\Documents\CarLoadDocumentItem.cs" />
     <Compile Include="Domain\Documents\CarUnloadDocument.cs" />
